Cast the colorID to an int and restrict it to 0-255

+++ Structs/GameDataTypes.cs
@@ -95,7 +95,7 @@ namespace ImpostersOrdeal
         public class Trainer : INamedEntity
         {
             public int trainerTypeID;
-            public int colorID;
+            public byte colorID;
             public byte fightType;
             public int arenaID;
             public int effectID;
